dsp.* from the Dolphin repository

+++ dsp.h
@@ -6,9 +6,9 @@
 #define ISQRT2 0.70710678118654f
 
 void dct_quantize(uint8_t *in_data, uint8_t *prediction, uint32_t width, uint32_t height,
-		int16_t *out_data, uint8_t *quantization);
+		int16_t *out_data, float *quantization);
 
 void dequantize_idct(int16_t *in_data, uint8_t *prediction, uint32_t width, uint32_t height,
-		uint8_t *out_data, uint8_t *quantization);
+		uint8_t *out_data, float *quantization);
 
 #endif  /* C63_DSP_SIMD_H_ */
